@@ -1,13 +1,15 @@
 name: flutter_sms_example
 description: Demonstrates how to use the flutter_sms plugin.
-publish_to: 'none'
+publish_to: "none"
 
 environment:
-  sdk: '>=2.12.0-259.8.beta <3.0.0'
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+
+  permission_handler: ^12.0.0+1
 
 dev_dependencies:
   flutter_sms:
@@ -17,4 +19,3 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
-
